rename flows[sequence,parallel] to just sequence,parallel
