+++ cmd/controller/main.go
@@ -25,9 +25,9 @@ import (
 	"knative.dev/eventing/pkg/reconciler/broker"
 	"knative.dev/eventing/pkg/reconciler/channel"
 	"knative.dev/eventing/pkg/reconciler/eventtype"
-	flowsparallel "knative.dev/eventing/pkg/reconciler/flowsparallel"
-	flowssequence "knative.dev/eventing/pkg/reconciler/flowssequence"
 	"knative.dev/eventing/pkg/reconciler/namespace"
+	parallel "knative.dev/eventing/pkg/reconciler/parallel"
+	sequence "knative.dev/eventing/pkg/reconciler/sequence"
 	"knative.dev/eventing/pkg/reconciler/subscription"
 	"knative.dev/eventing/pkg/reconciler/trigger"
 )
@@ -40,7 +40,7 @@ func main() {
 		trigger.NewController,
 		broker.NewController,
 		eventtype.NewController,
-		flowsparallel.NewController,
-		flowssequence.NewController,
+		parallel.NewController,
+		sequence.NewController,
 	)
 }
